% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab_hour}
\alias{grab_hour}
\title{Grab the hour only}
\usage{
grab_hour(x)
}
\arguments{
\item{x}{A numeric value to be grabed.}
}
\value{
The hour only.
}
\description{
Simply grab the hour from the time column.
}
