% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab_minute}
\alias{grab_minute}
\title{Grab the minute only}
\usage{
grab_minute(x)
}
\arguments{
\item{x}{A numeric value to be grabed.}
}
\value{
The minute only.
}
\description{
It help us to grab only minute form the datetime column
}
